Sources/requirements.txt
numpy~=1.18.1
pandas~=1.0.3
pymrmre~=1.0.0
scikit-learn~=0.22.1
pytorch~=1.4.0
